VF-166 Move Tab-component to component folder
